requirements.txt
alembic==0.9.9
blinker==1.4
click==6.7
dominate==2.3.1
extensions==0.4
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
numpy==1.14.4
pandas==0.23.0
PyJWT==1.6.4
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.7
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1

